test refactors (#6715)
