+++ conans/test/integration/only_source_test.py
@@ -5,7 +5,6 @@ from conans.model.ref import ConanFileReference
 from conans.paths import CONANFILE
 from conans.test.utils.cpp_test_files import cpp_hello_conan_files
 from conans.test.utils.tools import TestClient, TestServer
-from conans.util.files import load
 
 
 class OnlySourceTest(unittest.TestCase):
